week5/exercise13/bookstore_security/src/test/java/com/example/bookstore_security/controller/AuthControllerTest.java
package com.example.bookstore_security.controller;

import com.example.bookstore_security.dto.LoginRequest;
import com.example.bookstore_security.dto.RegisterRequest;
import com.example.bookstore_security.model.User;
import com.example.bookstore_security.security.JwtTokenProvider;
import com.example.bookstore_security.service.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.security.core.Authentication;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.hamcrest.Matchers.is;

@WebMvcTest(AuthController.class)
public class AuthControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @MockBean
    private JwtTokenProvider jwtTokenProvider;

    @InjectMocks
    private AuthController authController;

    @Mock
    private Authentication authentication;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    @WithMockUser(username = "user", roles = {"USER"})
    public void testLoginUser() throws Exception {
        // Arrange
        String token = "mocked-jwt-token";
        LoginRequest loginRequest = new LoginRequest();
        loginRequest.setUsername("user");
        loginRequest.setPassword("password");

        when(userService.loadUserByUsername(loginRequest.getUsername())).thenReturn(new User());
        when(jwtTokenProvider.createToken(any())).thenReturn(token);
        when(authentication.getName()).thenReturn("user");

        mockMvc.perform(MockMvcRequestBuilders.post("/api/auth/login")
                .contentType(MediaType.APPLICATION_JSON)
                .content("{\"username\":\"user\", \"password\":\"password\"}"))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.token", is(token)));
    }

    @Test
    public void testRegisterUser() throws Exception {
       
        RegisterRequest registerRequest = new RegisterRequest();
        registerRequest.setUsername("sa");
        registerRequest.setPassword("password");

       
        mockMvc.perform(MockMvcRequestBuilders.post("/api/auth/register")
                .contentType(MediaType.APPLICATION_JSON)
                .content("{\"username\":\"newuser\", \"password\":\"password\"}"))
                .andExpect(status().isOk());
    }
}
